[omni] Read and use touch slop multiplier for LPNH.
1) This part is missing from the previous set of CLs. `getLPNHCustomSlopMultiplier` is defined but unused. Similar as reading LPNHDuration, we should also read the touch slot multiplier in Launcher for LPNH.
2) Also, change type to float, which is consistent with our existing implementation and the definition in OverviewProxyService.
3) Slop value is in percentage and Launcher divide it by 100. We haven't got AGSA implementation yet so we can predefine this part.

Bug: 330444720
Bug: 330446188
Test: manual
Flag: legacy CUSTOM_LPNH_THRESHOLDS disabled

+++ quickstep/src/com/android/quickstep/util/AssistStateManager.java
@@ -58,7 +58,7 @@ public class AssistStateManager implements ResourceBasedOverride {
     }
 
     /** Get the Launcher overridden long press touch slop multiplier to trigger Assistant. */
-    public Optional<Long> getLPNHCustomSlopMultiplier() {
+    public Optional<Float> getLPNHCustomSlopMultiplier() {
         return Optional.empty();
     }
